refactor: decompose SubAgentRunner into focused modules
- Extract filter.py: SUBAGENT_EXCLUDED_TOOLS + filter_subagent_tools
- Extract formatter.py: _format_elapsed + build_origin_suffix
- Extract profiler.py: SubAgentProfiler (profile resolution, skill/tool filtering)
- Extract notifier.py: SubAgentNotifier (notification formatting and delivery)
- Extract executor.py: SubAgentExecutor (execution loop, timeout, error handling)
- Rewrite runner.py as thin facade (~300 lines public API + orchestration)
- Split test_subagent_runner.py (1,558 lines) into 4 focused test files
- Update test_subagent_progress.py imports for moved functions
- All public API signatures preserved for backward compatibility
- 2,890 tests passing, ruff clean
