+++ openpaw/runtime/subagent/__init__.py
@@ -1,5 +1,10 @@
 """Sub-agent runtime services."""
 
+from openpaw.runtime.subagent.filter import SUBAGENT_EXCLUDED_TOOLS, filter_subagent_tools
 from openpaw.runtime.subagent.runner import SubAgentRunner
 
-__all__ = ["SubAgentRunner"]
+__all__ = [
+    "SubAgentRunner",
+    "filter_subagent_tools",
+    "SUBAGENT_EXCLUDED_TOOLS",
+]
